Split branch creation logic out of Bump::PullRequestCreator
Previously, we had two classes for creating / updating pull requests:
- Bump::PullRequestCreator
- Bump::PullRequestUpdater

Both knew how to create commits for a set of dependency files, and the former
knew how to create PRs.

This PR re-jigs that setup into the following split:
- Bump::DependencyFilePusher
- Bump::PullRequestCreator

The former knows how to create commits and branches for a set of dependency
files. The later knows how to create a PR for a given branch. This is a better
split because:
- less logic is duplicated (previously both classes had their own, unshared
  logic for creating commits)
- an end-user might was to create branches without creating PRs. For example if
  they wanted to wait for specs to run before creating the PRs