Remove path-based dependencies from Gemfile
We currently don't support path-based dependencies, so, we
remove them from the Gemfile entirely.

We'll resolve the Gemfile without a line (it'll be removed)
from the file, which is not ideal but it'll at
least allow us to resolve what needs to update as opposed
to not resolving anything at all.